python/soracc_legacy/requirements.txt
pytest~=8.3.3
mypy~=1.10.0
mypy-extensions~=1.0.0
types-PyYAML~=6.0.12.20240917
bitstring~=4.2.3
PyYAML~=6.0.2
numpy~=1.26.4
safetensors~=0.4.5
bitarray~=2.9.3
tqdm~=4.66.5
types-tqdm~=4.67.0.20241221
matplotlib~=3.10.0
onnx~=1.17.0